% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeurl.R
\name{leg_info}
\alias{leg_info}
\title{Download legislative info related to legal codes}
\usage{
leg_info(codes)
}
\arguments{
\item{codes}{A vector of legal codes}
}
\description{
Download legislative info related to legal codes
}
